+++ src/Controller/UserController.php
@@ -5,10 +5,8 @@ namespace App\Controller;
 use App\Entity\Product;
 use App\Entity\User;
 use App\Form\ChangePasswordType;
+use App\Form\PasswordResetType;
 use App\Form\UserType;
-use App\Pagination\PaginatedCollection;
-use Pagerfanta\Doctrine\ORM\QueryAdapter;
-use Pagerfanta\Pagerfanta;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -18,10 +16,16 @@ class UserController extends ApiController
     /**
      * @Route("/users/{id}", name="users_get_item", methods="GET", requirements={"id"="\d+"})
      */
-    public function item(User $user)
+    public function item(int $id)
     {
-        $user = $this->serializer->serialize($user, 'json');
-        return new Response($user, Response::HTTP_OK, ['Content-Type' => 'application/json']);
+        $user = $this->entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+        if (!$user) {
+            throw $this->createNotFoundException(sprintf(
+                'No user found with id %d',
+                $id
+            ));
+        }
+        return $this->createApiResponse($user);
     }
 
     /**
@@ -29,59 +33,27 @@ class UserController extends ApiController
      */
     public function collection(Request $request)
     {
-
-        $page = $request->query->get('page', 1);
-        $qb = $this->entityManager->getRepository(User::class)->findAllQueryBuilder();
-        $adapter = new QueryAdapter($qb);
-        $pagerfanta = new Pagerfanta($adapter);
-        $pagerfanta->setCurrentPage($page);
-        $pagerfanta->getCurrentPageResults();
-        $users = [];
-        foreach ($pagerfanta->getCurrentPageResults() as $user) {
-            $users[] = $user;
-        }
-        $paginatedCollection = new PaginatedCollection(
-            $users,
-            $pagerfanta->getNbResults()
-        );
-        $route = 'users_get_collection';
-        $routeParams = [];
-        $createLinkUrl = function ($targetPage) use ($route, $routeParams) {
-          return $this->generateUrl($route, array_merge(
-             $routeParams,
-             ['page' => $targetPage]
-          ));
-        };
-        $paginatedCollection->addLink('self', $createLinkUrl($page));
-        $paginatedCollection->addLink('first', $createLinkUrl(1));
-        $paginatedCollection->addLink('last', $createLinkUrl($pagerfanta->getNbPages()));
-
-        if($pagerfanta->hasNextPage()) {
-            $paginatedCollection->addLink('next', $createLinkUrl($pagerfanta->getNextPage()));
-        }
-        if($pagerfanta->hasPreviousPage()) {
-            $paginatedCollection->addLink('prev', $createLinkUrl($pagerfanta->getNextPage()));
-        }
-
-        return $this->json(
-            $paginatedCollection->getResult('users'),
-            Response::HTTP_OK,
-            ['Content-Type' => 'application/json']
-        );
+        $filter = $request->query->get('filter');
+        $qb = $this->entityManager->getRepository(User::class)->findAllQueryBuilder($filter);
+        $paginatedCollection = $this->paginationFactory->createCollection($qb, $request, 'users_get_collection');
+        return $this->createApiResponse($paginatedCollection->getResult('users'));
     }
 
     /**
      * @Route("/users/{id}", name="users_delete_item", methods="DELETE", requirements={"id"="\d+"})
      */
-    public function delete(User $user)
+    public function delete(int $id)
     {
+        $user = $this->entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+        if (!$user) {
+            throw $this->createNotFoundException(sprintf(
+                'No user found with id %d',
+                $id
+            ));
+        }
         $this->entityManager->remove($user);
         $this->entityManager->flush();
-        return $this->json(
-            ['info' => 'User has been deleted.'],
-            Response::HTTP_OK,
-            ['Content-Type' => 'application/json']
-        );
+        return $this->createApiResponse(['info' => 'User has been deleted.']);
     }
 
     /**
@@ -89,13 +61,6 @@ class UserController extends ApiController
      */
     public function post(Request $request)
     {
-//        for($i = 0; $i < 25; $i++) {
-//            $user = new User();
-//            $user->setUsername('daniel'.$i);
-//            $user->setPassword('daniel'.$i);
-//            $user->setEmail('daniel'.$i.'@daniel.com');
-//            $this->entityManager->persist($user);
-//        }
         $data = $this->jsonDecode($request->getContent());
         $this->isJsonValid($data);
         $form = $this->createForm(UserType::class, new User());
@@ -104,24 +69,29 @@ class UserController extends ApiController
 
         /** @var User $user */
         $user = $form->getData();
+        $user->setConfirmationToken($this->tokenGenerator->getRandomSecureToken());
+        $user->setEnabled(false);
         $user->setPassword(
             $this->userPasswordEncoder->encodePassword($user, $user->getPassword())
         );
         $this->entityManager->persist($user);
         $this->entityManager->flush();
-
-        return $this->json(
-            ['info' => 'User has been added.'],
-            Response::HTTP_CREATED,
-            ['Content-Type' => 'application/json']
-        );
+        $this->mailer->sendConfirmationEmail($user);
+        return $this->createApiResponse(['info' => 'User has been added.'], Response::HTTP_CREATED);
     }
 
     /**
      * @Route("/users/{id}/change-password", name="users_change_password", methods="PATCH")
      */
-    public function changePassword(Request $request, User $user)
+    public function changePassword(Request $request, int $id)
     {
+        $user = $this->entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+        if (!$user) {
+            throw $this->createNotFoundException(sprintf(
+                'No user found with id %d',
+                $id
+            ));
+        }
         $data = $this->jsonDecode($request->getContent());
         $this->isJsonValid($data);
         $form = $this->createForm(ChangePasswordType::class, new User());
@@ -143,6 +113,29 @@ class UserController extends ApiController
 
         return $this->json(
             ['info' => 'Password has been changed.'],
+            Response::HTTP_OK,
+            ['Content-Type' => 'application/json']
+        );
+    }
+
+    /**
+     * @Route("/user-password-reset", name="app_password_reset", methods="POST")
+     */
+    public function userPasswordReset(Request $request)
+    {
+        $data = $this->jsonDecode($request->getContent());
+        $this->isJsonValid($data);
+        $form = $this->createForm(PasswordResetType::class, new User());
+        $form->submit($data);
+        $this->isFormValid($form);
+        $user = $this->entityManager->getRepository(User::class)->findOneBy(['email' => $form->getData()->getEmail()]);
+        if (!$user) {
+            throw $this->createNotFoundException(sprintf('Could not find user with e-mail %s', $form->getData()->getEmail()));
+        }
+        $user->setConfirmationToken($this->tokenGenerator->getRandomSecureToken());
+        $this->mailer->sendPasswordResetEmail($user);
+        return $this->json(
+            ['info' => 'E-mail with instructions to reset password has been sent.'],
             Response::HTTP_OK,
             ['Content-Type' => 'application/json']
         );
@@ -173,13 +166,17 @@ class UserController extends ApiController
      */
     public function userProducts(User $user)
     {
+        $user = $this->entityManager->getRepository(User::class)->findOneBy(['id' => $id]);
+        if (!$user) {
+            throw $this->createNotFoundException(sprintf(
+                'No user found with id %d',
+                $id
+            ));
+        }
         $products = $this->entityManager->getRepository(Product::class)->findBy(['owner' => $user]);
         // add form
         // add voter
-        return $this->json($products);
+        return $this->createApiResponse($products);
     }
-
-
-
 
 }
